Change encoding to UTF8

+++ common/client/message.cs
@@ -1,4 +1,4 @@
-﻿if (isObject(MessageFuncDict))
+if (isObject(MessageFuncDict))
 {
     MessageFuncDict.delete();
 }
